requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.3
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.18.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.4.0
packaging==25.0
parsel==1.10.0
Protego==0.4.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dotenv==1.1.0
python-telegram-bot==22.1
queuelib==1.8.0
redis==6.2.0
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
Scrapy==2.13.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
tldextract==5.3.0
Twisted==24.11.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
w3lib==2.3.1
websockets==15.0.1
zope.interface==7.2
